# contracts/openapi.json
{
  "components": {
    "schemas": {
      "AckIn": {
        "properties": {
          "ids": {
            "items": {
              "type": "integer"
            },
            "title": "Ids",
            "type": "array"
          }
        },
        "title": "AckIn",
        "type": "object"
      },
...
      "DeviceIn": {
        "properties": {
          "app_version": {
            "default": "",
            "maxLength": 32,
            "title": "App Version",
            "type": "string"
          },
          "model": {
            "default": "",
            "maxLength": 64,
            "title": "Model",
            "type": "string"
          },
          "platform": {
            "$ref": "#/components/schemas/Platform",
            "default": "android"
          },
          "push_token": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Push Token",
            "type": "string"
          }
        },
        "required": [
          "push_token"
        ],
        "title": "DeviceIn",
        "type": "object"
      },
      "DeviceRegistrert": {
        "description": "Svarmodell (se routers/auth.py).",
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "platform": {
            "title": "Platform",
            "type": "string"
          }
        },
        "required": [
          "id",
          "platform"
        ],
        "title": "DeviceRegistrert",
        "type": "object"
      },
...
      "FeedbackIn": {
        "properties": {
          "app_version": {
            "default": "",
            "maxLength": 32,
            "title": "App Version",
            "type": "string"
          },
          "body": {
            "maxLength": 10000,
            "minLength": 1,
            "title": "Body",
            "type": "string"
          },
          "device_model": {
            "default": "",
            "maxLength": 64,
            "title": "Device Model",
            "type": "string"
          },
          "subject": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Subject",
            "type": "string"
          },
          "user_id": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          }
        },
        "required": [
          "subject",
          "body"
        ],
        "title": "FeedbackIn",
        "type": "object"
      },
      "FeedbackMottatt": {
        "description": "Svarmodell (se routers/auth.py). 202, ikke 200: meldingen er lagret, men\nvideresendingen skjer i bakgrunnen og kan feile uten at brukeren merker\ndet - `forward_error` settes paa raden.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "status": {
            "examples": [
              "mottatt"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status"
        ],
        "title": "FeedbackMottatt",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Platform": {
        "enum": [
          "android",
          "ios"
        ],
        "title": "Platform",
        "type": "string"
      },
...
      "UnregisterIn": {
        "properties": {
          "push_token": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Push Token",
            "type": "string"
          }
        },
        "required": [
          "push_token"
        ],
        "title": "UnregisterIn",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Bestefar backend",
    "version": "0.2"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/devices": {
      "get": {
        "description": "Til «innlogget paa disse enhetene» i UI-et. Selve push-tokenet returneres\nIKKE - det gir ingen nytte i klienten, og et svar er et sted det kan lekke.",
        "operationId": "list_devices_v1_devices_get",
        "parameters": [
          {
            "description": "«Bearer <access-token>» (§1).",
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "«Bearer <access-token>» (§1).",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response List Devices V1 Devices Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Devices",
        "tags": [
          "enheter"
        ]
      },
      "put": {
        "description": "Idempotent med vilje - derfor PUT og ikke POST. Klienten kaller dette ved\nhver oppstart, og skal ikke trenge aa vite om tokenet er nytt.\n\n`push_token` er unik paa tvers av brukere. Byttes telefonen mellom to\nkontoer, skal raden FOELGE tokenet: ellers ville varsler til den forrige\nbrukeren havnet paa en telefon som naa er logget inn som noen andre.",
        "operationId": "register_device_v1_devices_put",
        "parameters": [
          {
            "description": "«Bearer <access-token>» (§1).",
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "«Bearer <access-token>» (§1).",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceRegistrert"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Device",
        "tags": [
          "enheter"
        ]
      }
    },
    "/v1/devices/unregister": {
      "post": {
        "description": "Kalles ved utlogging. Idempotent: et ukjent token gir ogsaa 204, av samme\ngrunn som /v1/auth/logout - avregistrering skal aldri feile og etterlate\nklienten i tro om at den fortsatt faar varsler.\n\nBare egne enheter kan avregistreres. Uten det filteret kunne hvem som helst\nskrudd av varslene til en annen ved aa gjette et token.",
        "operationId": "unregister_device_v1_devices_unregister_post",
        "parameters": [
          {
            "description": "«Bearer <access-token>» (§1).",
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "«Bearer <access-token>» (§1).",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnregisterIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unregister Device",
        "tags": [
          "enheter"
        ]
      }
    },
...
    "/v1/feedback": {
      "post": {
        "operationId": "submit_feedback_v1_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackMottatt"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Feedback",
        "tags": [
          "feedback"
        ]
      }
    },
...
    "/v1/messages/ack": {
      "post": {
        "description": "Klienten kvitterer naar meldingen er VIST. Serveren sletter ikke raden -\nden markeres som levert, saa en klient som krasjer mellom henting og\nvisning ikke mister meldingen for godt.",
        "operationId": "ack_messages_v1_messages_ack_post",
        "parameters": [
          {
            "description": "«Bearer <access-token>» (§1).",
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "«Bearer <access-token>» (§1).",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AckIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ack Messages",
        "tags": [
          "meldinger"
        ]
      }
    },
...
  }
}